+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - `Innmind\Filesystem\File\Content` is now a final class and its different implementations are declared internal, use the `Content` named constructors instead
 - `Innmind\Filesystem\Directory` no longer extends `File`, all previous function typed against `File` are now typed `File|Directory`
 - `Innmind\Filesystem\File` is now a final class instead of an interface
+- `Innmind\Filesystem\Directory` is now a final class instead of an interface
 
 ### Removed